+++ hosts/zaneyos-oem/variables.nix
@@ -9,13 +9,10 @@
   # SDDM background is set with stylixImage
   displayManager = "sddm";
 
-  # Themed tmux
+  # Emable/disable bundled applications
   tmuxEnable = true;
-
-  # Enable Alacritty
   alacrittyEnable = true;
-
-  # Enable ghostty
+  weztermEnable = true;
   ghosttyEnable = true;
 
   # P10K or starship prompt